Improve modal and add basic restart feature
