+++ src/components/Header.svelte
@@ -1,5 +1,10 @@
+<script lang="ts">
+    export let startNewGame: () => void
+</script>
+
 <header class="shadow-lg flex justify-between items-center p-4 bg-stone-800">
-    <!-- <svg
+    <svg
+        on:click={startNewGame}
         xmlns="http://www.w3.org/2000/svg"
         class="h-6 w-6"
         fill="none"
@@ -12,8 +17,7 @@
             stroke-width="2"
             d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
         />
-    </svg> -->
-    <p class="w-6 h-6">{" "}</p>
+    </svg>
     <h1 class="font-semibold text-4xl -mt-2">Iso Conquest</h1>
     <a href="https://github.com/Greenheart/iso-conquest">
         <svg
