+++ src/components/Modal.svelte
@@ -1,19 +1,32 @@
-<script lang="ts">
-    import { scale } from "svelte/transition"
-    let modal: HTMLDivElement
+<script lang="ts" context="module">
+    import Modal from "$components/BaseModal.svelte"
 
-    const classes = [
-        $$props.class ?? "",
-        "absolute left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2 shadow-2xl bg-stone-700",
-    ].join(" ")
+    interface ModalAction {
+        onClick: () => void
+        label: string
+    }
 </script>
 
-<div
-    class="fixed top-0 left-0 w-full h-full bg-black bg-opacity-30 pointer-events-none"
-/>
+<script lang="ts">
+    export let title: string
+    export let actions: ModalAction[]
+</script>
 
-<div class={classes} role="dialog" aria-modal="true" bind:this={modal} in:scale>
-    <slot name="header" />
-    <slot />
-    <slot name="footer" />
-</div>
+<Modal class="bg-stone-700 p-8 max-w-md w-full">
+    <h2 slot="header" class="font-semibold text-2xl text-center">
+        {title}
+    </h2>
+    <div class="my-8">
+        <slot />
+    </div>
+    <div slot="footer" class="flex justify-center flex-col">
+        <!-- svelte-ignore a11y-autofocus -->
+        {#each actions as { onClick, label }}
+            <button
+                class="px-12 py-4 bg-emerald-500 text-white hover:transform-gpu hover:scale-105 duration-100 text-xl font-semibold"
+                autofocus
+                on:click={onClick}>{label}</button
+            >
+        {/each}
+    </div>
+</Modal>
