+++ src/routes/__layout.svelte
@@ -1,10 +1,7 @@
 <script>
-    import Header from "$components/Header.svelte"
-
     import "../app.css"
 </script>
 
 <main class="text-emerald-500">
-    <Header />
     <slot />
 </main>
